+++ app/assets/javascripts/Results/main.js
@@ -1,6 +1,3 @@
-// UI LIBRARIES
-//= require DropDownMenu/DropDownMenu
-
 // Source code highlighting
 //= require Annotations/globals
 //= require Annotations/annotation_manager
@@ -16,4 +13,3 @@
 // Page-specific Javascript
 //= require Grader/marking
 //= require panes
-//= require Results/keybinding