+++ app/controllers/course_summaries_controller.rb
@@ -53,7 +53,7 @@ class CourseSummariesController < ApplicationController
     marking_schemes = current_role.student? ? MarkingScheme.none : current_course.marking_schemes.order(id: :asc)
     intervals = 20
     table_data = marking_schemes.map { |m| { data: m.students_grade_distribution(current_role, intervals) } }
-    labels = (0..intervals - 1).map { |i| "#{5 * i}-#{5 * i + 5}" }
+    labels = (0..(intervals - 1)).map { |i| "#{5 * i}-#{5 * i + 5}" }
 
     summary = marking_schemes.map do |m|
       grades = m.students_weighted_grades_array(current_role)