+++ app/javascript/Components/instructor_table.jsx
@@ -2,26 +2,65 @@ import React from "react";
 import {createRoot} from "react-dom/client";
 import PropTypes from "prop-types";
 
-import ReactTable from "react-table";
-import {selectFilter} from "./Helpers/table_helpers";
+import Table from "./table/table";
+import {createColumnHelper} from "@tanstack/react-table";
+import {faPencil} from "@fortawesome/free-solid-svg-icons";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 
 class InstructorTable extends React.Component {
   constructor() {
     super();
     this.state = {
       data: [],
-      counts: {all: 0, active: 0, inactive: 0},
       loading: true,
     };
     this.fetchData = this.fetchData.bind(this);
+
+    const columnHelper = createColumnHelper();
+    this.columns = [
+      columnHelper.accessor("user_name", {
+        header: () => I18n.t("activerecord.attributes.user.user_name"),
+      }),
+      columnHelper.accessor("first_name", {
+        header: () => I18n.t("activerecord.attributes.user.first_name"),
+      }),
+      columnHelper.accessor("last_name", {
+        header: () => I18n.t("activerecord.attributes.user.last_name"),
+      }),
+      columnHelper.accessor("email", {
+        header: () => I18n.t("activerecord.attributes.user.email"),
+      }),
+      columnHelper.accessor("hidden", {
+        header: () => I18n.t("roles.active") + "?",
+        filterFn: "equalsString",
+        meta: {
+          filterVariant: "select",
+        },
+      }),
+      columnHelper.accessor("id", {
+        id: "id",
+        enableSorting: false,
+        enableColumnFilter: false,
+        header: () => I18n.t("actions"),
+        cell: props => (
+          <a
+            href={Routes.edit_course_instructor_path(this.props.course_id, props.getValue())}
+            aria-label={I18n.t("edit")}
+            title={I18n.t("edit")}
+          >
+            <FontAwesomeIcon icon={faPencil} />
+          </a>
+        ),
+      }),
+    ];
   }
 
   componentDidMount() {
-    this.fetchData();
+    this.fetchData().then(data => this.setState({data: this.processData(data)}));
   }
 
   fetchData() {
-    fetch(Routes.course_instructors_path(this.props.course_id), {
+    return fetch(Routes.course_instructors_path(this.props.course_id), {
       headers: {
         Accept: "application/json",
       },
@@ -31,75 +70,19 @@ class InstructorTable extends React.Component {
           return response.json();
         }
       })
-      .then(res => {
-        this.setState({data: res.data, counts: res.counts, loading: false});
-      });
+      .then(json => json.data);
+  }
+
+  processData(data) {
+    data.forEach(row => (row.hidden = I18n.t(row.hidden ? "roles.inactive" : "roles.active")));
+    return data;
   }
 
   render() {
-    const {data} = this.state;
     return (
-      <ReactTable
-        data={data}
-        columns={[
-          {
-            Header: I18n.t("activerecord.attributes.user.user_name"),
-            accessor: "user_name",
-          },
-          {
-            Header: I18n.t("activerecord.attributes.user.first_name"),
-            accessor: "first_name",
-          },
-          {
-            Header: I18n.t("activerecord.attributes.user.last_name"),
-            accessor: "last_name",
-          },
-          {
-            Header: I18n.t("activerecord.attributes.user.email"),
-            accessor: "email",
-          },
-          {
-            Header: I18n.t("roles.active") + "?",
-            accessor: "hidden",
-            Cell: ({value}) => (value ? I18n.t("roles.inactive") : I18n.t("roles.active")),
-            filterMethod: (filter, row) => {
-              if (filter.value === "all") {
-                return true;
-              } else {
-                return (
-                  (filter.value === "active" && !row[filter.id]) ||
-                  (filter.value === "inactive" && row[filter.id])
-                );
-              }
-            },
-            Filter: selectFilter,
-            filterOptions: [
-              {
-                value: "active",
-                text: `${I18n.t("roles.active")} (${this.state.counts.active})`,
-              },
-              {
-                value: "inactive",
-                text: `${I18n.t("roles.inactive")} (${this.state.counts.inactive})`,
-              },
-            ],
-            filterAllOptionText: `${I18n.t("all")} (${this.state.counts.all})`,
-          },
-          {
-            Header: I18n.t("actions"),
-            accessor: "id",
-            Cell: data => (
-              <span>
-                <a href={Routes.edit_course_instructor_path(this.props.course_id, data.value)}>
-                  {I18n.t("edit")}
-                </a>
-              </span>
-            ),
-            sortable: false,
-          },
-        ]}
-        filterable
-        loading={this.state.loading}
+      <Table
+        data={this.state.data}
+        columns={this.columns}
         noDataText={I18n.t("instructors.empty_table")}
       />
     );